+++ versions/1.21.1/src/main/java/me/pajic/zombieimprovements/mixin/LivingEntityRendererMixin.java
@@ -5,7 +5,7 @@ import com.llamalad7.mixinextras.sugar.Local;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import me.pajic.zombieimprovements.Main;
-import me.pajic.zombieimprovements.util.ZombieData;
+import me.pajic.zombieimprovements.util.ZombieExtension;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
@@ -43,11 +43,11 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, M extend
             @Local(ordinal = 4) float netHeadYaw,
             @Local(ordinal = 5) float headPitch
     ) {
-        if (Main.CONFIG.leaderRedAura.get() && entity instanceof Zombie zombie && zombie.getEntityData().get(ZombieData.LEADER)) {
+        if (Main.CONFIG.leaderRedAura.get() && entity instanceof Zombie zombie && ((ZombieExtension) zombie).zi$isLeader()) {
             float f = zombie.tickCount + partialTick;
             EntityModel<T> entityModel = getModel();
             entityModel.prepareMobModel(entity, limbSwing, limbSwingAmount, partialTick);
-            VertexConsumer vertexConsumer = bufferSource.getBuffer(RenderType.energySwirl(ZombieData.ZOMBIE_POWER_LAYER, (f * 0.01F) % 1.0F, f * 0.01F % 1.0F));
+            VertexConsumer vertexConsumer = bufferSource.getBuffer(RenderType.energySwirl(Main.ZOMBIE_POWER_LAYER, (f * 0.01F) % 1.0F, f * 0.01F % 1.0F));
             entityModel.setupAnim(entity, limbSwing, limbSwingAmount, getBob(entity, partialTick), netHeadYaw, headPitch);
             entityModel.renderToBuffer(poseStack, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY, -8355712);
         }