+++ versions/1.21.8/src/main/java/me/pajic/zombieimprovements/mixin/LivingEntityRendererMixin.java
@@ -6,7 +6,6 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import me.pajic.zombieimprovements.Main;
 import me.pajic.zombieimprovements.util.HumanoidRenderStateExtension;
-import me.pajic.zombieimprovements.util.ZombieData;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
@@ -41,7 +40,7 @@ public abstract class LivingEntityRendererMixin<T extends LivingEntity, S extend
         if (Main.CONFIG.leaderRedAura.get() && renderState.entityType.is(EntityTypeTags.ZOMBIES) && ((HumanoidRenderStateExtension) renderState).zi$isLeader()) {
             float f = renderState.ageInTicks;
             M entityModel = getModel();
-            VertexConsumer vertexConsumer = bufferSource.getBuffer(RenderType.energySwirl(ZombieData.ZOMBIE_POWER_LAYER, (f * 0.01F) % 1.0F, f * 0.01F % 1.0F));
+            VertexConsumer vertexConsumer = bufferSource.getBuffer(RenderType.energySwirl(Main.ZOMBIE_POWER_LAYER, (f * 0.01F) % 1.0F, f * 0.01F % 1.0F));
             entityModel.setupAnim(renderState);
             entityModel.renderToBuffer(poseStack, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY, -8355712);
         }